chore(lint): task-28.3 clippy 存量修到 -D warnings 全绿（surgical 等价，0 行为变更）
实测 clippy 存量 ~33 小 lint（cargo clippy --workspace --all-targets）→ cargo clippy --fix
机械修 + 手动收尾：field_reassign_with_default→结构字面量 / doc_lazy_continuation→文档缩进 /
ptr_arg &PathBuf·&mut Vec→slice / while_let_loop / slice::from_ref / vec_init_then_push→vec!。
2 处 targeted allow：core/src/lib.rs pb/pb_console 生成代码 clippy::all；events.rs EventBus::send
result_large_err（返 tokio broadcast SendError，boxing ripple 公共签名 → 刻意不 box + 注释）。

复验：cargo clippy --workspace --all-targets -- -D warnings exit 0；cargo test --workspace 全过
（contextforge-core lib 187 passed，各 integration 0 failed，bench 7 passed）。不 blanket 抑制。

+++ core/src/data_plane/search.rs
@@ -561,9 +561,9 @@ fn seconds_to_iso(unix_secs: i64) -> String {
     // Civil-date arithmetic (no chrono dep). Days since 1970-01-01.
     let days = unix_secs.div_euclid(86_400);
     let secs_of_day = unix_secs.rem_euclid(86_400);
-    let hour = (secs_of_day / 3600) as i64;
-    let minute = ((secs_of_day % 3600) / 60) as i64;
-    let second = (secs_of_day % 60) as i64;
+    let hour = secs_of_day / 3600;
+    let minute = (secs_of_day % 3600) / 60;
+    let second = secs_of_day % 60;
     // Convert days since epoch to civil date (Howard Hinnant algorithm).
     let z = days + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
